another_surveys/project/telegram-bot/requirements.txt
python-telegram-bot==13.7
kafka-python==2.0.2
pymongo==3.11.3
prometheus-client==0.14.1
